// src/lib/client/types/ElectionChoice.ts
import type { ElectionCampaignModel } from './ElectionCampaignModel.ts'
import type { ElectionCandidateModel } from './ElectionCandidateModel.ts'

export type ElectionChoice = {
  /**
   * @type string | undefined
   */
  _id?: string
  /**
   * @type object
   */
  candidate: ElectionCandidateModel
  running_mate?: ElectionCandidateModel | null
  campaign?: ElectionCampaignModel | null
}